made output fairly consistent with Aug 10 - 09baa05
Difference (at least one of them) was apparently in the mesh. Needed to do a zeroQ projection with the standalone mesh for the triangulation, but the non-zeroQ pair otherwise.. There's still a slight difference in the data, and I'm not sure what it's from. get_ocubo.m produces identical results; it could be sphconvhulln, get_octpairs, get_interp, or proj_up. I think those are the only places it could be realistically.