calculate dataprops alongside propList instead of in interp5DOF.m, due to bug
There seems to be a major bug with get_octpairs (or GBoct2five.m). In generating an octochoric octonion and symmetrizing, the disorientation of the new 5DOF parameters are the same, but the boundary plane angle differs (and not just by an inversion as one might expect for an interior point). For example:
                      nA(1)  nA(2)  nA(3)
before get_octpairs: -0.6505 0.4207 0.6324
after get_octpairs:  -0.8550 0.5172 0.0377
This suggests the GB is actually being changed as a result of the get_octpairs operation. Could this have to do with the "active rotation" in the octonion conversion, or with the analytic derivative in the distance formula?